[add]トークルームを作成si,
